pyproject.toml
[project]
name = "xjj-housekeeper"
version = "0.1.0"
description = "视频文件管理系统"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
keywords = ["video", "file-management", "metadata", "filename-formatter"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dynamic = ["dependencies"]

[tool.poetry]
packages = [
    {include = "tools"}
]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-mock = "^3.12.0"
pytest-xdist = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# 系统依赖说明
# 以下系统依赖需要手动安装：
# - FFmpeg (包含 ffprobe): 用于视频元数据提取
#   macOS: brew install ffmpeg
#   Ubuntu/Debian: sudo apt-get install ffmpeg
#   Windows: 下载并安装 FFmpeg 二进制文件
# - SQLite3: 通常系统自带，Python 内置支持
